+++ Tests/Test_Old/Test_SetK_PlanetBuild/test_a_build_star.py
@@ -6,7 +6,7 @@ from scipy.constants import G
 import TidalPy
 
 
-from TidalPy import world_config_loc
+from TidalPy.paths import get_worlds_dir
 from TidalPy.structures import build_world
 
 cancri_star_config = {
@@ -57,7 +57,7 @@ def test_build_star_from_manual_config():
 
 def test_build_star_from_file_loaded_config():
     # Test loading a star from a user-provided configuration file
-    cancri_filepath = os.path.join(world_config_loc, '55cnc.json')
+    cancri_filepath = os.path.join(get_worlds_dir(), '55cnc.json')
     with open(cancri_filepath, 'r') as cancri_config_file:
         cancri_star = build_world('55-Cancri', cancri_config_file)