+++ TidalPy/structures/world_builder/config_handler.py
@@ -2,7 +2,7 @@ import copy
 
 import json5
 
-from TidalPy import world_config_loc
+from TidalPy.paths import get_worlds_dir
 from TidalPy.utilities.io.pathing import get_all_files_of_type
 
 from TidalPy.logger import get_logger
@@ -69,7 +69,7 @@ def check_for_duplicate_worlds(world_configs: dict):
 
 # Find all planet configurations and import their config files
 # Locate all planet configurations
-known_worlds_files = get_all_files_of_type(world_config_loc, ['cfg', 'json', 'json5'])
+known_worlds_files = get_all_files_of_type(get_worlds_dir(), ['cfg', 'json', 'json5'])
 known_worlds_cfg = dict()
 check_for_duplicate_worlds(known_worlds_cfg)
 _configs_loaded = False